Move vault client methods into separate files
